Fix more issues old code..
