+++ itachallenge-score/src/main/java/com/itachallenge/score/component/ExecutionResult.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 @Setter
 public class ExecutionResult {
     private ExecutionResultDto executionResultDto;
-    private String executionResult;
+    private String executionResultMsg;
 }
