+++ itachallenge-score/src/test/java/com/itachallenge/score/filter/AsciiFilterTest.java
@@ -7,16 +7,39 @@ import static org.junit.Assert.assertTrue;
 
 class AsciiFilterTest {
 
+    String ASCIIvalid = """
+                  import java.util.Arrays;
+                    
+            public class Main {
+                public static void main(String[] args) {
+                    String numbers = "3,1,4,1,5,9";
+                    int[] numArray = Arrays.stream(numbers.split(",")).mapToInt(Integer::parseInt).toArray();
+                    Arrays.sort(numArray);
+                    System.out.println(Arrays.toString(numArray));
+                }
+            }
+                """;
+
+    String ASCIIinvalid = """
+                              import java.util.Arrays;
+                    
+            public class βain {
+                public static void main(String[] args) {
+                    String numbers = "©, 3,1,4,1,5,9";
+                    int[] numArray = Arrays.stream(numbers.split(",")).mapToInt(Integer::parseInt).toArray();
+                    Arrays.sort(numArray);
+                    System.out.println(Arrays.toString(numArray)); """;
+
     @Test
     void testFilterValid() {
         AsciiFilter filter = new AsciiFilter();
-        assertTrue("The String had a valid ASCII chars", filter.apply(StringTest.ASCIIvalid));
+        assertTrue("The String had a valid ASCII chars", filter.apply(ASCIIvalid));
     }
 
     @Test
     void testFilterInvalid() {
         AsciiFilter filter = new AsciiFilter();
-        assertFalse("The String had invalid ASCII chars", filter.apply(StringTest.ASCIIinvalid));
+        assertFalse("The String had invalid ASCII chars", filter.apply(ASCIIinvalid));
     }
 
     @Test
